use size_t for filesizes (off_t?)

+++ FormatConverter/Hdf5Reader.h
@@ -29,7 +29,7 @@ protected:
 	void finish( ) override;
 
 	ReadResult fill( SignalSet& data, const ReadResult& lastfill ) override;
-	int getSize( const std::string& input ) const override;
+	size_t getSize( const std::string& input ) const override;
 
 private:
 	Hdf5Reader( const Hdf5Reader& );
